[Godfile execution-engine step 1](1) extract task-runner-git
Relocate the git and gh subprocess plumbing out of task-runner.ts into a
host-free sibling, task-runner-git.ts. The relocation is verbatim except for
one hardening: the createMergeWorktree/cloneMergeWorktree clone and fetch
calls gain a `--` end-of-options separator before their caller-derived
remote, refspec, and clone source, so git cannot read a value beginning with
`-` as an option (e.g. --upload-pack=<cmd>). This resolves the CodeQL
js/second-order-command-line-injection alerts on the new file with no
behavior change for legitimate refs/URLs.

TaskRunner keeps one-line delegates so the MergeRunnerHost and
ConflictResolverHost method obligations stay untouched.
