+++ packages/execution-engine/src/task-runner.ts
@@ -42,7 +42,6 @@ import {
   ensureLocalBranchForMerge,
   collectTransitiveNonMergeTaskIds,
 } from './merge-runner.js';
-import { normalizeBranchForGithubCli } from './github-branch-ref.js';
 import {
   resolveConflictImpl,
   fixWithAgentImpl,
@@ -63,9 +62,8 @@ import {
   validateReviewStackPrBody,
   type PrAuthoringContext,
 } from './pr-authoring.js';
-import { assertNotGitConfigMutation, ensureRemoteUrl } from './git-config-mutation.js';
-import { killProcessGroup, SIGKILL_TIMEOUT_MS } from './process-utils.js';
-import { retryTransientGitHubCli } from './git-utils.js';
+import { ensureRemoteUrl } from './git-config-mutation.js';
+import * as gitPlumbing from './task-runner-git.js';
 import { PRE_START_HEARTBEAT_INTERVAL_MS, nextLeaseExpiry } from './task-runner-launch-support.js';
 import { buildWorkRequest } from './task-runner-prepare.js';
 import { dispatchExecutor } from './task-runner-dispatch.js';
@@ -75,10 +73,6 @@ export type { TaskHeartbeatEvent, TaskRunnerCallbacks } from './task-runner-call
 type ReviewGateState = NonNullable<TaskState['execution']['reviewGate']>;
 type ReviewGateArtifact = ReviewGateState['artifacts'][number];
 type ReviewGateArtifactStatus = ReviewGateArtifact['status'];
-
-
-const DEFAULT_GIT_OPERATION_TIMEOUT_MS = 15 * 60 * 1000;
-const GITHUB_TARGET_REPO_ENV = 'INVOKER_GITHUB_TARGET_REPO';
 
 export type ActiveExecutionHandle = ExecutorHandle & { attemptId?: string };
 export type ActiveExecutionEntry = {
@@ -157,70 +151,6 @@ export interface ReviewGateCiFailureTrigger {
 
 export interface ReviewGateCiFailureLifecyclePublisher {
   publish(trigger: ReviewGateCiFailureTrigger): void | Promise<void>;
-}
-
-function getGitOperationTimeoutMs(): number {
-  const raw = process.env.INVOKER_GIT_NETWORK_TIMEOUT_MS?.trim();
-  if (raw === '0') return 0;
-  if (!raw) return DEFAULT_GIT_OPERATION_TIMEOUT_MS;
-  const parsed = Number.parseInt(raw, 10);
-  if (!Number.isFinite(parsed) || parsed < 0) return DEFAULT_GIT_OPERATION_TIMEOUT_MS;
-  return parsed;
-}
-
-function execGitWithTimeout(args: string[], cwd: string): Promise<string> {
-  return new Promise((resolvePromise, reject) => {
-    const child = spawn('git', args, {
-      cwd,
-      stdio: ['ignore', 'pipe', 'pipe'],
-      detached: process.platform !== 'win32',
-    });
-    const timeoutMs = getGitOperationTimeoutMs();
-    let stdout = '';
-    let stderr = '';
-    let settled = false;
-    let timeout: ReturnType<typeof setTimeout> | undefined;
-
-    const finish = (fn: () => void): void => {
-      if (settled) return;
-      settled = true;
-      if (timeout) clearTimeout(timeout);
-      fn();
-    };
-
-    if (timeoutMs > 0) {
-      timeout = setTimeout(() => {
-        killProcessGroup(child, 'SIGTERM');
-        const forceKill = setTimeout(() => {
-          killProcessGroup(child, 'SIGKILL');
-        }, SIGKILL_TIMEOUT_MS);
-        forceKill.unref?.();
-        finish(() => reject(new Error(
-          `git ${args.join(' ')} exceeded git operation timeout (${timeoutMs}ms) in ${cwd}. ` +
-          'Set INVOKER_GIT_NETWORK_TIMEOUT_MS to adjust (0 = unbounded).',
-        )));
-      }, timeoutMs);
-      timeout.unref?.();
-    }
-
-    child.stdout?.on('data', (d: Buffer) => { stdout += d.toString(); });
-    child.stderr?.on('data', (d: Buffer) => { stderr += d.toString(); });
-    child.on('error', (err) => {
-      finish(() => reject(new Error(`Failed to spawn git: ${err.message}`)));
-    });
-    child.on('close', (code, signal) => {
-      finish(() => {
-        if (code === 0) {
-          resolvePromise(stdout.trim());
-          return;
-        }
-        reject(new Error(
-          `git ${args.join(' ')} failed (code ${code ?? 'null'}${signal ? ` signal ${signal}` : ''}): ` +
-          `${stderr.trim()}${stdout.trim() ? '\n' + stdout.trim() : ''}`,
-        ));
-      });
-    });
-  });
 }
 
 const NOOP_LOGGER: Logger = {
@@ -1618,295 +1548,47 @@ export class TaskRunner {
    * @internal Read-only git queries only. Config mutations must use git-config-mutation helpers.
    */
   execGitReadonly(args: string[], cwd?: string): Promise<string> {
-    assertNotGitConfigMutation(args, 'TaskRunner.execGitReadonly');
-    return execGitWithTimeout(args, cwd ?? this.cwd);
+    return gitPlumbing.execGitReadonly(args, cwd ?? this.cwd);
   }
 
   /** @internal */ execGitIn(args: string[], dir: string): Promise<string> {
-    assertNotGitConfigMutation(args, 'TaskRunner.execGitIn');
-    return execGitWithTimeout(args, dir);
+    return gitPlumbing.execGitIn(args, dir);
   }
 
-  /** @internal */ async createMergeWorktree(ref: string, label: string, repoUrl?: string): Promise<string> {
-    const invokerHomeRoot = process.env.INVOKER_DB_DIR
-      ? resolve(process.env.INVOKER_DB_DIR)
-      : resolve(homedir(), '.invoker');
-    const mergeCloneRoot = resolve(invokerHomeRoot, 'merge-clones');
-    mkdirSync(mergeCloneRoot, { recursive: true });
-    const clonePath = mkdtempSync(resolve(mergeCloneRoot, `${label}-`));
-
-    // Determine clone source: prefer pool mirror (has latest remote refs), fall back to host repo
-    let cloneSource: string = this.cwd;
-    let originUrl: string | undefined;
-    if (repoUrl) {
-      const mirrorPath = await this.ensureRepoMirrorPath(repoUrl);
-      if (mirrorPath) {
-        cloneSource = mirrorPath;
-        originUrl = repoUrl;
-      } else {
-        this.logger.warn(`[createMergeWorktree] Pool mirror unavailable for ${repoUrl}, falling back to host repo`);
-      }
-    }
-
-    await this.cloneMergeWorktree(cloneSource, clonePath);
-    // Detach HEAD so the fetch can overwrite all branch refs (including the default branch)
-    const headSha = (await this.execGitIn(['rev-parse', 'HEAD'], clonePath)).trim();
-    await this.execGitIn(['update-ref', '--no-deref', 'HEAD', headSha], clonePath);
-    // Mirror all branches as local refs so bare branch names resolve.
-    await this.execGitIn(['fetch', 'origin', '+refs/heads/*:refs/heads/*'], clonePath);
-
-    // Reconfigure origin to the real remote URL (GitHub) so subsequent push/fetch
-    // operations go directly to GitHub, bypassing any intermediate clone.
-    if (!originUrl) {
-      // Fallback: read origin from the host repo (old behavior)
-      originUrl = (await this.execGitReadonly(['remote', 'get-url', 'origin'])).trim();
-    }
-    await ensureRemoteUrl({
-      cwd: clonePath,
-      remote: 'origin',
-      url: originUrl,
-      context: { caller: 'TaskRunner.createMergeWorktree', detail: `${label}:${ref}` },
+  /** @internal */ createMergeWorktree(ref: string, label: string, repoUrl?: string): Promise<string> {
+    return gitPlumbing.createMergeWorktree(ref, label, repoUrl, {
+      cwd: this.cwd,
+      logger: this.logger,
+      ensureRepoMirrorPath: (url) => this.ensureRepoMirrorPath(url),
     });
-
-    // Refresh the requested base branch from the real remote. The pool mirror's
-    // local refs/heads/* can go stale after force-pushes or history rewrites,
-    // causing merge conflicts when experiment branches are based on the new
-    // history but the clone got the old branch tip from the pool.
-    const normalizedRef = ref.trim();
-    const strippedRemoteRef = normalizeBranchForGithubCli(normalizedRef);
-    const remoteName = 'origin';
-    const baseRef = normalizedRef.startsWith('origin/')
-      ? normalizedRef.slice('origin/'.length)
-      : strippedRemoteRef;
-    try {
-      await this.execGitIn(
-        ['fetch', remoteName, `+refs/heads/${baseRef}:refs/remotes/${remoteName}/${baseRef}`],
-        clonePath,
-      );
-    } catch {
-      // Non-critical: pool's ref may still be valid
-    }
-
-    // Resolve ref in the clone (not in host repo — the clone has mirrored branches).
-    // Accept both "feature/x" and "origin/feature/x" forms, and tolerate missing
-    // origin tracking refs for local-only stacked branches.
-    const tryResolve = async (expr: string): Promise<string | undefined> => {
-      try {
-        return (await this.execGitIn(['rev-parse', '--verify', `${expr}^{commit}`], clonePath)).trim();
-      } catch {
-        return undefined;
-      }
-    };
-
-    const candidates = Array.from(new Set([
-      `refs/remotes/${remoteName}/${baseRef}`,
-      `${remoteName}/${baseRef}`,
-      normalizedRef,
-      strippedRemoteRef,
-      `refs/heads/${strippedRemoteRef}`,
-    ]));
-
-    let refSha: string | undefined;
-    for (const candidate of candidates) {
-      refSha = await tryResolve(candidate);
-      if (refSha) break;
-    }
-
-    if (!refSha) {
-      // Last chance: fetch only the requested branch from origin, then retry.
-      // This can happen if clone source had stale refs at submit time.
-      try {
-        await this.execGitIn(
-          ['fetch', remoteName, `+refs/heads/${strippedRemoteRef}:refs/remotes/${remoteName}/${strippedRemoteRef}`],
-          clonePath,
-        );
-      } catch {
-        // Best-effort; keep error message from final resolve below.
-      }
-      for (const candidate of candidates) {
-        refSha = await tryResolve(candidate);
-        if (refSha) break;
-      }
-    }
-
-    // Fallback: if the requested ref is one of the common default branch names
-    // and it doesn't exist, try the alternate (main↔master). Plan YAML files
-    // sometimes specify "main" for repos whose default branch is "master" or
-    // vice versa.
-    if (!refSha) {
-      const alternates: Record<string, string> = { main: 'master', master: 'main' };
-      const alt = alternates[strippedRemoteRef];
-      if (alt) {
-        try {
-          await this.execGitIn(
-            ['fetch', remoteName, `+refs/heads/${alt}:refs/remotes/${remoteName}/${alt}`],
-            clonePath,
-          );
-        } catch {
-          // Best-effort
-        }
-        const altCandidates = [
-          `refs/remotes/${remoteName}/${alt}`,
-          `${remoteName}/${alt}`,
-          alt,
-          `refs/heads/${alt}`,
-        ];
-        for (const candidate of altCandidates) {
-          refSha = await tryResolve(candidate);
-          if (refSha) break;
-        }
-      }
-    }
-
-    if (!refSha) {
-      throw new Error(
-        `Branch "${ref}" required by the merge/gate step was not found on the remote (${originUrl}). ` +
-        `This branch is retrieved from origin, but it is not there — it was never pushed to origin, ` +
-        `or it has since been deleted. Push the branch to origin (or point the workflow's base at a ` +
-        `branch that exists on origin), then rerun the gate. Recreating the task alone will not help ` +
-        `while the branch is missing from origin. ` +
-        `(resolved in clone ${clonePath}; tried ${candidates.join(', ')})`,
-      );
-    }
-    await this.execGitIn(['checkout', '--detach', refSha], clonePath);
-    return clonePath;
   }
 
-  /** @internal */ async cloneMergeWorktree(cloneSource: string, clonePath: string): Promise<void> {
-    try {
-      // Hard-linked objects make local pool clones near-instant while keeping refs isolated.
-      await this.execGitReadonly(['clone', '--local', '--no-checkout', cloneSource, clonePath]);
-    } catch (err) {
-      const message = err instanceof Error ? `${err.message}\n${err.stack ?? ''}` : String(err);
-      if (!message.includes('Invalid cross-device link') && !message.includes('EXDEV')) {
-        throw err;
-      }
-
-      // CI may place the repo/mirror and temp merge clone on different mounts,
-      // where Git's hardlink-based local clone fails with EXDEV.
-      this.logger.warn(
-        `[createMergeWorktree] Local clone crossed filesystems; retrying without hardlinks: ${message.split('\n')[0]}`,
-      );
-      rmSync(clonePath, { recursive: true, force: true });
-      await this.execGitReadonly(['clone', '--no-local', '--no-checkout', cloneSource, clonePath]);
-    }
+  /** @internal */ cloneMergeWorktree(cloneSource: string, clonePath: string): Promise<void> {
+    return gitPlumbing.cloneMergeWorktree(cloneSource, clonePath, (args) => this.execGitReadonly(args), this.logger);
   }
 
-  /** @internal */ async removeMergeWorktree(dir: string): Promise<void> {
-    try {
-      rmSync(dir, { recursive: true, force: true });
-    } catch (err) {
-      this.logger.warn('[TaskRunner] removeMergeWorktree failed (best-effort)', {
-        dir,
-        error: err instanceof Error ? err.message : String(err),
-        err,
-      });
-    }
+  /** @internal */ removeMergeWorktree(dir: string): Promise<void> {
+    return gitPlumbing.removeMergeWorktree(dir, this.logger);
   }
 
-  async detectDefaultBranch(): Promise<string> {
-    try {
-      const ref = await this.execGitReadonly(['symbolic-ref', 'refs/remotes/origin/HEAD']);
-      return ref.replace('refs/remotes/origin/', '');
-    } catch {
-      try {
-        await this.execGitReadonly(['rev-parse', '--verify', 'main']);
-        return 'main';
-      } catch {
-        return 'master';
-      }
-    }
+  detectDefaultBranch(): Promise<string> {
+    return gitPlumbing.detectDefaultBranch((args) => this.execGitReadonly(args));
   }
 
   /** @internal */ execGh(args: string[], cwd?: string): Promise<string> {
-    const effectiveCwd = cwd ?? this.cwd;
-    return new Promise((resolvePromise, reject) => {
-      const child = spawn('gh', args, {
-        cwd: effectiveCwd,
-        stdio: ['ignore', 'pipe', 'pipe'],
-      });
-      let stdout = '';
-      let stderr = '';
-      child.stdout?.on('data', (d: Buffer) => { stdout += d.toString(); });
-      child.stderr?.on('data', (d: Buffer) => { stderr += d.toString(); });
-      child.on('close', (code) => {
-        if (code === 0) resolvePromise(stdout.trim());
-        else reject(new Error(`gh ${args[0]} ${args[1]} failed (code ${code}): ${stderr.trim()}`));
-      });
-    });
+    return gitPlumbing.execGh(args, cwd ?? this.cwd);
   }
 
-  /** @internal */ async execPr(baseBranch: string, featureBranch: string, title: string, body?: string, cwd?: string): Promise<string> {
-    const ghBase = normalizeBranchForGithubCli(baseBranch);
-    const ghHead = normalizeBranchForGithubCli(featureBranch);
-    const effectiveCwd = cwd ?? this.cwd;
-    const targetRepo = await this.resolveGithubTargetRepo(effectiveCwd);
-    const repoOwner = targetRepo.split('/')[0];
-
-    const listOutput = await retryTransientGitHubCli(() => this.execGh([
-      'api', `repos/${targetRepo}/pulls`,
-      '--method', 'GET',
-      '-f', `head=${repoOwner}:${ghHead}`,
-      '-f', 'state=open',
-      '-f', 'per_page=1',
-    ], effectiveCwd));
-
-    const existing: Array<{ html_url?: string; url?: string; number: number }> = JSON.parse(listOutput || '[]');
-    if (existing.length > 0) {
-      const pr = existing[0];
-      const editArgs = [
-        'api', `repos/${targetRepo}/pulls/${pr.number}`,
-        '--method', 'PATCH',
-        '-f', `base=${ghBase}`,
-        '-f', `title=${title}`,
-      ];
-      if (body) editArgs.push('-f', `body=${body}`);
-      await retryTransientGitHubCli(() => this.execGh(editArgs, effectiveCwd));
-      return pr.html_url ?? pr.url ?? '';
-    }
-
-    const createOutput = await this.execGh([
-      'api', `repos/${targetRepo}/pulls`,
-      '--method', 'POST',
-      '-f', `base=${ghBase}`,
-      '-f', `head=${ghHead}`,
-      '-f', `title=${title}`,
-      '-f', `body=${body ?? ''}`,
-    ], effectiveCwd);
-    try {
-      const pr = JSON.parse(createOutput) as { html_url?: string; url?: string };
-      return pr.html_url ?? pr.url ?? createOutput;
-    } catch {
-      return createOutput;
-    }
-  }
-
-  private async resolveGithubTargetRepo(cwd: string): Promise<string> {
-    const explicitTarget = process.env[GITHUB_TARGET_REPO_ENV]?.trim();
-    if (explicitTarget) {
-      if (/^[^/\s]+\/[^/\s]+$/.test(explicitTarget)) return explicitTarget;
-      throw new Error(
-        `Invalid ${GITHUB_TARGET_REPO_ENV}="${explicitTarget}". Expected format "owner/repo".`,
-      );
-    }
-
-    try {
-      const url = await this.execGitIn(['remote', 'get-url', 'origin'], cwd);
-      const parsed = this.parseGitHubRepoNwo(url);
-      if (parsed) return parsed;
-    } catch {
-      // fall through
-    }
-
-    throw new Error(
-      'Unable to resolve GitHub target repo. ' +
-      `Set ${GITHUB_TARGET_REPO_ENV}=owner/repo or configure a parseable origin GitHub remote.`,
+  /** @internal */ execPr(baseBranch: string, featureBranch: string, title: string, body?: string, cwd?: string): Promise<string> {
+    return gitPlumbing.execPr(
+      baseBranch,
+      featureBranch,
+      title,
+      body,
+      cwd ?? this.cwd,
+      (args, ghCwd) => this.execGh(args, ghCwd),
+      (args, dir) => this.execGitIn(args, dir),
     );
-  }
-
-  private parseGitHubRepoNwo(url: string): string | undefined {
-    const m = url.trim().match(/github\.com[:/]([^/]+\/[^/.]+?)(?:\.git)?\/?$/i);
-    return m?.[1];
   }
 
   // ── Experiment Branch Merging ────────────────────────────
@@ -2911,39 +2593,10 @@ export class TaskRunner {
   }
 
   /** @internal */ gitLogMessage(commitHash: string, cwd?: string): Promise<string> {
-    return new Promise((resolvePromise, reject) => {
-      const child = spawn('git', ['log', '-1', '--format=%B', commitHash], {
-        cwd: cwd ?? this.cwd,
-        stdio: ['ignore', 'pipe', 'pipe'],
-      });
-      let stdout = '';
-      child.stdout?.on('data', (d: Buffer) => { stdout += d.toString(); });
-      child.on('error', (err) => {
-        reject(new Error(`Failed to spawn git: ${err.message}`));
-      });
-      child.on('close', (code) => {
-        if (code === 0) resolvePromise(stdout.trim());
-        else reject(new Error(`git log failed (code ${code})`));
-      });
-    });
+    return gitPlumbing.gitLogMessage(commitHash, cwd ?? this.cwd);
   }
 
   /** @internal */ gitDiffStat(branch: string, cwd?: string): Promise<string> {
-    return new Promise((resolvePromise, reject) => {
-      const baseBranch = this.defaultBranch ?? 'master';
-      const child = spawn('git', ['diff', '--stat', '--stat-count=20', `${baseBranch}...${branch}`], {
-        cwd: cwd ?? this.cwd,
-        stdio: ['ignore', 'pipe', 'pipe'],
-      });
-      let stdout = '';
-      child.stdout?.on('data', (d: Buffer) => { stdout += d.toString(); });
-      child.on('error', (err) => {
-        reject(new Error(`Failed to spawn git: ${err.message}`));
-      });
-      child.on('close', (code) => {
-        if (code === 0) resolvePromise(stdout.trim());
-        else reject(new Error(`git diff --stat failed (code ${code})`));
-      });
-    });
+    return gitPlumbing.gitDiffStat(branch, this.defaultBranch, cwd ?? this.cwd);
   }
 }
